groot/requirements.txt
ComradeAI
pika
Pillow
python-dotenv
requests
debugpy